requirements.txt
pyTelegramBotAPI==3.7.3
psycopg2-binary==2.8.6
python-dotenv==0.15.0